Core/Spells: Fix magical spell damage to be neither very low or completely null - Derived function under Player class had uint16_t variable instead of virtual function's uint32 variable - Simplify strings in WowPlayer.h

+++ src/world/Units/Creatures/Creature.cpp
@@ -2008,7 +2008,7 @@ Group* Creature::GetGroup()
     return NULL;
 }
 
-int32 Creature::GetDamageDoneMod(uint32 school)
+int32 Creature::GetDamageDoneMod(uint16_t school)
 {
     if (school >= SCHOOL_COUNT)
         return 0;
@@ -2016,7 +2016,7 @@ int32 Creature::GetDamageDoneMod(uint32 school)
     return ModDamageDone[ school ];
 }
 
-float Creature::GetDamageDonePctMod(uint32 school)
+float Creature::GetDamageDonePctMod(uint16_t school)
 {
     if (school >= SCHOOL_COUNT)
         return 0;
